cluster/operation: Prune tomestone instances also need to destroy monitor
