+++ pkg/cluster/operation/action.go
@@ -18,11 +18,9 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cluster/api"
 	"github.com/pingcap/tiup/pkg/cluster/module"
 	"github.com/pingcap/tiup/pkg/cluster/spec"
 	"github.com/pingcap/tiup/pkg/logger/log"
@@ -164,208 +162,6 @@ func NeedCheckTombstone(topo *spec.Specification) bool {
 	}
 
 	return false
-}
-
-// DestroyTombstone remove the tombstone node in spec and destroy them.
-// If returNodesOnly is true, it will only return the node id that can be destroy.
-func DestroyTombstone(
-	getter ExecutorGetter,
-	cluster *spec.Specification,
-	returNodesOnly bool,
-	options Options,
-	tlsCfg *tls.Config,
-) (nodes []string, err error) {
-	return DestroyClusterTombstone(getter, cluster, returNodesOnly, options, tlsCfg)
-}
-
-// DestroyClusterTombstone remove the tombstone node in spec and destroy them.
-// If returNodesOnly is true, it will only return the node id that can be destroy.
-func DestroyClusterTombstone(
-	getter ExecutorGetter,
-	cluster *spec.Specification,
-	returNodesOnly bool,
-	options Options,
-	tlsCfg *tls.Config,
-) (nodes []string, err error) {
-	var pdClient = api.NewPDClient(cluster.GetPDList(), 10*time.Second, tlsCfg)
-
-	binlogClient, err := api.NewBinlogClient(cluster.GetPDList(), tlsCfg)
-	if err != nil {
-		return nil, errors.AddStack(err)
-	}
-
-	filterID := func(instance []spec.Instance, id string) (res []spec.Instance) {
-		for _, ins := range instance {
-			if ins.ID() == id {
-				res = append(res, ins)
-			}
-		}
-		return
-	}
-
-	var kvServers []spec.TiKVSpec
-	for _, s := range cluster.TiKVServers {
-		if !s.Offline {
-			kvServers = append(kvServers, s)
-			continue
-		}
-
-		id := s.Host + ":" + strconv.Itoa(s.Port)
-
-		tombstone, err := pdClient.IsTombStone(id)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		if !tombstone {
-			kvServers = append(kvServers, s)
-			continue
-		}
-
-		nodes = append(nodes, id)
-		if returNodesOnly {
-			continue
-		}
-
-		instances := (&spec.TiKVComponent{Specification: cluster}).Instances()
-		instances = filterID(instances, id)
-
-		err = StopComponent(getter, instances, options.OptTimeout)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		err = DestroyComponent(getter, instances, cluster, options)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-	}
-
-	var flashServers []spec.TiFlashSpec
-	for _, s := range cluster.TiFlashServers {
-		if !s.Offline {
-			flashServers = append(flashServers, s)
-			continue
-		}
-
-		id := s.Host + ":" + strconv.Itoa(s.FlashServicePort)
-
-		tombstone, err := pdClient.IsTombStone(id)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		if !tombstone {
-			flashServers = append(flashServers, s)
-			continue
-		}
-
-		nodes = append(nodes, id)
-		if returNodesOnly {
-			continue
-		}
-
-		instances := (&spec.TiFlashComponent{Specification: cluster}).Instances()
-		id = s.Host + ":" + strconv.Itoa(s.GetMainPort())
-		instances = filterID(instances, id)
-
-		err = StopComponent(getter, instances, options.OptTimeout)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		err = DestroyComponent(getter, instances, cluster, options)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-	}
-
-	var pumpServers []spec.PumpSpec
-	for _, s := range cluster.PumpServers {
-		if !s.Offline {
-			pumpServers = append(pumpServers, s)
-			continue
-		}
-
-		id := s.Host + ":" + strconv.Itoa(s.Port)
-
-		tombstone, err := binlogClient.IsPumpTombstone(id)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		if !tombstone {
-			pumpServers = append(pumpServers, s)
-		}
-
-		nodes = append(nodes, id)
-		if returNodesOnly {
-			continue
-		}
-
-		instances := (&spec.PumpComponent{Specification: cluster}).Instances()
-		instances = filterID(instances, id)
-		err = StopComponent(getter, instances, options.OptTimeout)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		err = DestroyComponent(getter, instances, cluster, options)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-	}
-
-	var drainerServers []spec.DrainerSpec
-	for _, s := range cluster.Drainers {
-		if !s.Offline {
-			drainerServers = append(drainerServers, s)
-			continue
-		}
-
-		id := s.Host + ":" + strconv.Itoa(s.Port)
-
-		tombstone, err := binlogClient.IsDrainerTombstone(id)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		if !tombstone {
-			drainerServers = append(drainerServers, s)
-		}
-
-		nodes = append(nodes, id)
-		if returNodesOnly {
-			continue
-		}
-
-		instances := (&spec.DrainerComponent{Specification: cluster}).Instances()
-		instances = filterID(instances, id)
-
-		err = StopComponent(getter, instances, options.OptTimeout)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-
-		err = DestroyComponent(getter, instances, cluster, options)
-		if err != nil {
-			return nil, errors.AddStack(err)
-		}
-	}
-
-	if returNodesOnly {
-		return
-	}
-
-	cluster.TiKVServers = kvServers
-	cluster.TiFlashServers = flashServers
-	cluster.PumpServers = pumpServers
-	cluster.Drainers = drainerServers
-
-	return
 }
 
 // Restart the cluster.
